+++ publish-cli/src/main/java/com/ericsson/eiffel/remrem/publish/cli/CliOptions.java
@@ -1,6 +1,21 @@
+/*
+    Copyright 2017 Ericsson AB.
+    For a full list of individual contributors, please see the commit history.
+    Licensed under the Apache License, Version 2.0 (the "License");
+    you may not use this file except in compliance with the License.
+    You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+    Unless required by applicable law or agreed to in writing, software
+    distributed under the License is distributed on an "AS IS" BASIS,
+    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+    See the License for the specific language governing permissions and
+    limitations under the License.
+*/
 package com.ericsson.eiffel.remrem.publish.cli;
 
 import java.util.ArrayList;
+import java.util.Map;
 
 import org.apache.commons.cli.CommandLine;
 import org.apache.commons.cli.CommandLineParser;
@@ -13,6 +28,7 @@ import org.apache.commons.cli.Options;
 import org.apache.commons.lang3.ArrayUtils;
 
 import com.ericsson.eiffel.remrem.publish.config.PropertiesConfig;
+import com.ericsson.eiffel.remrem.shared.VersionService;
 
 public class CliOptions {
 	
@@ -63,6 +79,7 @@ public class CliOptions {
         options.addOption("mp", "messaging_protocol", true, "name of messaging protocol to be used, e.g. eiffel3, eiffelsemantics, default is eiffelsemantics");
         options.addOption("domain", "domainId", true, "identifies the domain that produces the event");
         options.addOption("ud", "user_domain_suffix", true, "user domain suffix");
+        options.addOption("v", "lists the versions of publish and all loaded protocols");
         contentGroup = createContentGroup();
         options.addOptionGroup(contentGroup);
     }
@@ -110,7 +127,9 @@ public class CliOptions {
         if (commandLine.hasOption("h")) {
     	    System.out.println("You passed help flag.");
     	    help(0);
-        } else {
+        } else if (commandLine.hasOption("v")) {
+            printVersions();
+        }else {
             checkRequiredOptions();
         }
     }
@@ -249,5 +268,29 @@ public class CliOptions {
         addErrorCode(errorCode);
       else
         System.exit(errorCode);
+    }
+    
+    /**
+     * Lists the versions of publish and all loaded protocols  
+     */
+    @SuppressWarnings({ "rawtypes", "unchecked" })
+    private static void printVersions() {
+        Map versions = new VersionService().getMessagingVersions();
+        Map<String, String> endpointVersions = (Map<String, String>) versions.get("endpointVersions");
+        Map<String, String> serviceVersion = (Map<String, String>) versions.get("serviceVersion");
+
+        if(serviceVersion != null) {
+            System.out.print("REMReM Publish version ");
+            for (String version: serviceVersion.values()) {
+                System.out.println(version);
+            }
+        }
+        if(endpointVersions != null) {
+            System.out.println("Available endpoints");
+            for (Map.Entry<String, String> entry : endpointVersions.entrySet()) {
+                System.out.println(entry);
+            }
+        }
+        exit(0);
     }
 }